+++ NEWS.md
@@ -3,6 +3,7 @@
 calculation and downloading of several tiles
 * update cog_2022 dataset to cog_2023
 * remove `get_apicarto_commune()` now supported by `get_apicarto_cadastre()`
+* `jsonlite` is used instead of `geojsonsf` package
 
 # happign 0.1.9
 * add `get_apicarto_viticole()`